requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
tqdm
matplotlib
sqlalchemy
psycopg2-binary